-Minor fix in UG for all-bases mode -Make minConfidenceScore in VariantEval a double so non-integer values can be used (requested by Steve H).
git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@2290 348d0f76-0448-11de-a6fe-93d51630548a

+++ java/src/org/broadinstitute/sting/playground/gatk/walkers/varianteval/VariantEvalWalker.java
@@ -30,7 +30,7 @@ import java.util.*;
 public class VariantEvalWalker extends RefWalker<Integer, Integer> {
 //public class VariantEvalWalker extends RodWalker<Integer, Integer> {
     @Argument(shortName="minPhredConfidenceScore", doc="Minimum confidence score to consider an evaluation SNP a variant", required=false)
-    public int minConfidenceScore = -1;
+    public double minConfidenceScore = -1.0;
 
     @Argument(shortName="printVariants", doc="If true, prints the variants in all of the variant tracks that are examined", required=false)
     public boolean printVariants = false;
